tinygen-backend/pyproject.toml
[project]
name = "tinygen-backend"
version = "0.1.0"
description = "TinyGen backend with Modal serverless functions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "modal>=1.1.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.4",
    "supabase>=2.18.0",
]
